backend/requirements.txt

flask
flask-cors
joblib
newspaper3k
tensorflow>=2.6.0
numpy
scikit-learn
gunicorn
